Extract photon_feedback_derive.rs from turn.rs (Phase 8 batch 1, #688)
Issue #688 (parent #680, Phase 8) batch 1 — opener PR. Create the new
sibling module `src/agent/loop_run/photon_feedback_derive.rs` and
move the photon feedback derive core (types + Case F predicate +
outcome_detail const) out of `turn.rs`.

Skipping Phases 6 & 7: the `PlanWriteSummary` / `PlanWriteSectionDelta`
types (Phase 6 / #686) are already housed in `actor_loop_flow.rs` (moved
during earlier Phase 1 / "Split plan write summary helpers" commits);
and the `RepairValidationTargetState` / `PreparedRepairIntentEdits<'a>`
/ `AppliedRepairCandidate` types (Phase 7 / #687) are already housed in
`verifier_orchestration.rs` (moved during Phase 2 work). Both phases are
effectively complete at the module level — the original meta-issue #680
goal was extraction out of turn.rs, not creation of yet another sibling.

Migrated to `photon_feedback_derive.rs` (`pub(crate)` for types,
`pub` for the const re-export target):

* `struct PhotonOutcomeInputs<'a>` (10 fields, all bounded — no
  free-text; audit boundary stays at type level)
* `struct PhotonFeedbackOutcome { outcome, outcome_detail }` (both
  `Option<&'static str>` — runtime data cannot leak into outbound
  payloads)
* `fn case_f_condition_met(&PhotonOutcomeInputs<'_>) -> bool` (the
  no-progress 4-condition AND SSOT)
* `pub const PHOTON_OUTCOME_DETAIL_NO_PROGRESS_DESPITE_INJECT: &str
    = "no_progress_despite_inject"` (Case F outcome_detail allowlist)

`loop_run.rs` adjustments:

* Registered `mod photon_feedback_derive;` with a Phase-8 / Issue-688
  doc comment.
* Updated the existing `pub use turn::PHOTON_OUTCOME_DETAIL_NO_PROGRESS_DESPITE_INJECT;`
  to `pub use photon_feedback_derive::PHOTON_OUTCOME_DETAIL_NO_PROGRESS_DESPITE_INJECT;`
  so `tests/photon_evaluate_signal_smoke.rs` keeps importing it via
  `anvil::agent::loop_run::PHOTON_OUTCOME_DETAIL_NO_PROGRESS_DESPITE_INJECT`
  without path changes (pre-existing re-export — not a new DR3-001
  exception).

`turn.rs` adjustments:

* Removed the struct + const definitions (originally at lines 545-657).
* Added top-of-file `use super::photon_feedback_derive::{
    PHOTON_OUTCOME_DETAIL_NO_PROGRESS_DESPITE_INJECT, PhotonFeedbackOutcome,
    PhotonOutcomeInputs, case_f_condition_met,
  };` so existing call sites resolve unchanged.

`CLAUDE.md` File Map updated: added `photon_feedback_derive.rs` entry.

LOC delta:

* `turn.rs`: 29,919 → 29,793 (-126 LOC)
* `photon_feedback_derive.rs`: 0 → 146 (new)
* Cumulative `turn.rs`: 39,489 → 29,793 (-9,696 LOC, -24.6%)

`pub(crate)` for the types (turn.rs internal), `pub` for the const
via the existing pre-Phase-8 re-export path. `turn.rs` is the only
in-crate consumer for the types.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
